feat: move theme mode switcher from chat to sidebar
